Bugs fixed. Added vehicle detail fetch function.

+++ Entities/Concrete/Car.cs
@@ -12,7 +12,7 @@ namespace Entities.Concrete
         public int SerialId { get; set; }
         public int ColorId { get; set; }
         public short ModelYear { get; set; }
-        public decimal Kilometer { get; set; }
+        public int Kilometer { get; set; }
         public decimal DailyPrice { get; set; }
         public string Description { get; set; }
